// PetClinicFrontend/src/app/model/Doctor.ts
export class Doctor {
  doctorId: number | undefined;
  firstName: string | undefined;
  lastName: string | undefined;
  email: string | undefined;
  password: string | undefined;
  phoneNumber: string | undefined;
  startScheduleTime: string | undefined;
  endScheduleTime: string | undefined;
  numberOfRatings: number | undefined;
  starsSum: number | undefined;
  rating: number | undefined;
}